Disentangle this a bit ...
Sequence#starts_at would usually be an alias of Expression#ts.

But Sequence#starts_at is overridden.

So Sequence#ts needs to be overridden, too, to use the local override of #starts_at.

There is no need to do this via the superclass Subexpression, though.

Subexpression#ts is already equal to Subexpression#starts_at, so overriding it there achieves nothing in that class itself.